app/backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-gis==1.2.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
